shipit_code_coverage: Move artifacts handling code in a separate module
