+++ src/shipit_code_coverage/shipit_code_coverage/codecov.py
@@ -5,16 +5,15 @@ import shutil
 import tarfile
 import requests
 import hglib
-from threading import Lock
 import concurrent.futures
 from concurrent.futures import ThreadPoolExecutor
 import sqlite3
-import time
 
 from cli_common.log import get_logger
 from cli_common.command import run_check
 
 from shipit_code_coverage import taskcluster, uploader
+from shipit_code_coverage.artifacts import ArtifactsHandler
 from shipit_code_coverage.github import GitHubUtils
 from shipit_code_coverage.notifier import Notifier
 from shipit_code_coverage.secrets import secrets
@@ -42,16 +41,16 @@ class CodeCov(object):
         self.access_token = access_token
 
         if revision is None:
-            self.task_ids = {
+            task_ids = {
                 'linux': taskcluster.get_last_task('linux'),
                 'windows': taskcluster.get_last_task('win'),
             }
 
-            task_data = taskcluster.get_task_details(self.task_ids['linux'])
+            task_data = taskcluster.get_task_details(task_ids['linux'])
             self.revision = task_data['payload']['env']['GECKO_HEAD_REV']
             self.from_pulse = False
         else:
-            self.task_ids = {
+            task_ids = {
                 'linux': taskcluster.get_task('mozilla-central', revision, 'linux'),
                 'windows': taskcluster.get_task('mozilla-central', revision, 'win'),
             }
@@ -60,128 +59,14 @@ class CodeCov(object):
             self.notifier = Notifier(revision, client_id, access_token)
 
         if self.from_pulse:
-            self.suites_to_ignore = ['awsy', 'talos']
+            suites_to_ignore = ['awsy', 'talos']
         else:
-            self.suites_to_ignore = []
+            suites_to_ignore = []
 
         logger.info('Mercurial revision', revision=self.revision)
 
+        self.artifactsHandler = ArtifactsHandler(task_ids, suites_to_ignore)
         self.githubUtils = GitHubUtils(cache_root, client_id, access_token)
-
-    def get_artifact_path(self, platform, chunk, artifact):
-        return 'ccov-artifacts/%s_%s_%s' % (platform, chunk, os.path.basename(artifact['name']))
-
-    def get_chunks(self):
-        return list(set([f.split('_')[1] for f in os.listdir('ccov-artifacts')]))
-
-    def get_coverage_artifacts(self, platform=None, suite=None, chunk=None):
-        files = os.listdir('ccov-artifacts')
-
-        if suite is not None and chunk is not None:
-            raise Exception('suite and chunk can\'t both have a value')
-
-        # Filter artifacts according to platform, suite and chunk.
-        filtered_files = []
-        for fname in files:
-            if platform is not None and not fname.startswith('%s_' % platform):
-                continue
-
-            if suite is not None and suite not in fname:
-                continue
-
-            if chunk is not None and ('%s_code-coverage' % chunk) not in fname:
-                continue
-
-            filtered_files.append('ccov-artifacts/' + fname)
-
-        return filtered_files
-
-    def download_coverage_artifacts(self):
-        mkdir('ccov-artifacts')
-
-        # The test tasks for the Linux and Windows builds are in the same group,
-        # but the following code is generic and supports build tasks split in
-        # separate groups.
-        groups = set([taskcluster.get_task_details(build_task_id)['taskGroupId'] for build_task_id in self.task_ids.values()])
-        test_tasks = [
-            task
-            for group in groups
-            for task in taskcluster.get_tasks_in_group(group)
-            if taskcluster.is_coverage_task(task)
-        ]
-
-        FINISHED_STATUSES = ['completed', 'failed', 'exception']
-        ALL_STATUSES = FINISHED_STATUSES + ['unscheduled', 'pending', 'running']
-        STATUS_VALUE = {
-            'exception': 1,
-            'failed': 2,
-            'completed': 3,
-        }
-
-        downloaded_tasks = {}
-        downloaded_tasks_lock = Lock()
-
-        def should_download(status, chunk_name, platform_name):
-            with downloaded_tasks_lock:
-                # If the chunk hasn't been downloaded before, this is obviously the best task
-                # to download it from.
-                if (chunk_name, platform_name) not in downloaded_tasks:
-                    download_lock = Lock()
-                    downloaded_tasks[(chunk_name, platform_name)] = {
-                        'status': status,
-                        'lock': download_lock,
-                    }
-                else:
-                    task = downloaded_tasks[(chunk_name, platform_name)]
-
-                    if STATUS_VALUE[status] > STATUS_VALUE[task['status']]:
-                        task['status'] = status
-                        download_lock = task['lock']
-                    else:
-                        return None
-
-                download_lock.acquire()
-                return download_lock
-
-        def download_artifact(test_task):
-            status = test_task['status']['state']
-            assert status in ALL_STATUSES
-            while status not in FINISHED_STATUSES:
-                time.sleep(60)
-                status = taskcluster.get_task_status(test_task['status']['taskId'])['status']['state']
-                assert status in ALL_STATUSES
-
-            chunk_name = taskcluster.get_chunk(test_task['task']['metadata']['name'])
-            platform_name = taskcluster.get_platform(test_task['task']['metadata']['name'])
-            # Ignore awsy and talos as they aren't actually suites of tests.
-            if any(to_ignore in chunk_name for to_ignore in self.suites_to_ignore):
-                return
-
-            # If we have already downloaded this chunk from another task, check if the
-            # other task has a better status than this one.
-            download_lock = should_download(status, chunk_name, platform_name)
-            if download_lock is None:
-                return
-
-            test_task_id = test_task['status']['taskId']
-            for artifact in taskcluster.get_task_artifacts(test_task_id):
-                if not any(n in artifact['name'] for n in ['code-coverage-grcov.zip', 'code-coverage-jsvm.zip']):
-                    continue
-
-                artifact_path = self.get_artifact_path(platform_name, chunk_name, artifact)
-                taskcluster.download_artifact(artifact_path, test_task_id, artifact['name'])
-                logger.info('%s artifact downloaded' % artifact_path)
-
-            download_lock.release()
-
-        def download_artifact_task(test_task):
-            return lambda: download_artifact(test_task)
-
-        with ThreadPoolExecutorResult() as executor:
-            for test_task in test_tasks:
-                executor.submit(download_artifact_task(test_task))
-
-        logger.info('Code coverage artifacts downloaded')
 
     def generate_info(self, commit_sha=None, platform=None, suite=None, chunk=None, out_format='coveralls', options=[]):
         cmd = [
@@ -210,7 +95,7 @@ class CodeCov(object):
             else:
                 cmd.extend(['--service-job-number', '1'])
 
-        cmd.extend(self.get_coverage_artifacts(platform, suite, chunk))
+        cmd.extend(self.artifactsHandler.get(platform, suite, chunk))
         cmd.extend(options)
 
         return run_check(cmd)
@@ -334,7 +219,7 @@ class CodeCov(object):
         with ThreadPoolExecutor(max_workers=4) as executor:
             futures = []
             for platform in ['linux', 'windows']:
-                for chunk in self.get_chunks():
+                for chunk in self.artifactsHandler.get_chunks():
                     futures.append(executor.submit(get_files_task(platform, chunk)))
 
             with sqlite3.connect('chunk_mapping.sqlite') as conn:
@@ -375,7 +260,7 @@ class CodeCov(object):
     def go(self):
         with ThreadPoolExecutorResult(max_workers=2) as executor:
             # Thread 1 - Download coverage artifacts.
-            executor.submit(lambda: self.download_coverage_artifacts())
+            executor.submit(lambda: self.artifactsHandler.download_all())
 
             # Thread 2 - Clone mozilla-central.
             executor.submit(lambda: self.clone_mozilla_central(self.revision))
